pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "duckclaw"
version = "0.1.0"
description = "Powerful AI — built for you, built with you, built securely. Local-first personal AI assistant."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "DuckClaw Contributors" }]
keywords = ["ai", "assistant", "local-llm", "privacy", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # API & Server
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "jinja2>=3.1.4",
    "python-multipart>=0.0.9",
    "websockets>=12.0",

    # LLM
    "litellm>=1.40.0",
    "anthropic>=0.30.0",

    # Memory
    "chromadb>=0.5.0",
    "aiosqlite>=0.20.0",

    # CLI
    "click>=8.1.7",
    "questionary>=2.0.1",
    "rich>=13.7.0",

    # Config
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.1",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",

    # Sprint 2: Messaging
    "python-telegram-bot>=21.0",
    "discord.py>=2.3.0",

    # Sprint 2: Skills
    "duckduckgo-search>=6.0.0",
    "mss>=9.0.0",
    "Pillow>=10.0.0",
    "apscheduler>=3.10.0",

    # Sprint 3: Browser (optional install via playwright install)
    "playwright>=1.40.0",
]

[project.optional-dependencies]
camera = [
    "opencv-python>=4.9.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
]

[project.scripts]
duckclaw = "duckclaw.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/duckclaw"
Repository = "https://github.com/yourusername/duckclaw"
Issues = "https://github.com/yourusername/duckclaw/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["duckclaw*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
